[Refactor] navigation 패키지 생성하여 navigation 관련 클래스 관리
